requirements_prod.txt
-r requirements_common.txt
psycopg2==2.6.1
uwsgi==2.0.12
django-redis-cache==1.6.5